wouter: docs for SoundDevice
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@597 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/AY8910.cc
@@ -208,7 +208,7 @@ void AY8910::wrtReg(byte reg, byte value)
 }
 
 
-void AY8910::setVolume(int newVolume)
+void AY8910::setVolume(short newVolume)
 {
 	// calculate the volume->voltage conversion table
 	// The AY-3-8910 has 16 levels, in a logarithmic scale (3dB per step)
